moved PublicNextcloudImage component to features/nextcloud_integration
