+++ frontend/src/features/nextcloud_integration/components/PublicNextcloudImage.tsx
